Clean: Merge STEX GaussBlur
