+++ src/dewarping/CMakeLists.txt
@@ -12,7 +12,6 @@ SET(
         DistortionModel.cpp DistortionModel.h
         DistortionModelBuilder.cpp DistortionModelBuilder.h
         RasterDewarper.cpp RasterDewarper.h
-        STEX_GaussBlur.cpp STEX_GaussBlur.h
         STEX_HomographicTransform.h
         STEX_ToPoint.h
         STEX_ToVec.h
